cloud-functions/bi-consumer/requirements.txt
google-cloud-bigquery==3.13.0
kafka-python==2.0.2
google-auth==2.23.4 